+++ packages/ricos-content/web/src/converters/draft/fromDraft/convertRicosPluginData.ts
@@ -22,6 +22,12 @@ import {
   Link_Target,
 } from 'ricos-schema';
 import { TO_RICOS_DATA } from './consts';
+import {
+  ComponentData,
+  FileComponentData,
+  ImageComponentData,
+  VideoComponentData,
+} from '../../../types';
 
 export const convertBlockDataToRicos = (blockType: string, data) => {
   const newData = cloneDeep(data);
@@ -52,12 +58,12 @@ export const convertBlockDataToRicos = (blockType: string, data) => {
   return fromJSON(newData);
 };
 
-const convertContainerData = data => {
-  const { size, alignment, width } = data.config;
-  let spoiler: PluginContainerData_Spoiler | undefined;
-  if (data.config.spoiler?.enabled) {
-    const { description, buttonContent } = data.config.spoiler;
-    spoiler = { description, buttonText: buttonContent };
+const convertContainerData = (data: { config?: ComponentData['config']; containerData }) => {
+  const { size, alignment, width, spoiler } = data.config || {};
+  let newSpoiler: PluginContainerData_Spoiler | undefined;
+  if (spoiler?.enabled) {
+    const { description, buttonContent } = spoiler;
+    newSpoiler = { description, buttonText: buttonContent };
   }
   const type =
     size && (size === 'inline' ? PluginContainerData_Width_Type.CUSTOM : kebabToConstantCase(size));
@@ -67,14 +73,19 @@ const convertContainerData = data => {
       customWidth: typeof width === 'number' ? width : undefined,
     },
     alignment: alignment && kebabToConstantCase(alignment),
-    spoiler,
+    spoiler: newSpoiler,
   };
 };
 
-const convertVideoData = data => {
+const convertVideoData = (data: {
+  src?: string | VideoComponentData;
+  metadata?: { thumbnail_url?: string; width?: number; height?: number };
+  video;
+  thumbnail;
+}) => {
   if (typeof data.src === 'string') {
     data.video = { src: { url: data.src } };
-    const { thumbnail_url, width, height } = data.metadata;
+    const { thumbnail_url, width, height } = data.metadata || {};
     data.thumbnail = {
       src: { url: thumbnail_url },
       width,
@@ -90,16 +101,31 @@ const convertVideoData = data => {
   }
 };
 
-const convertDividerData = data => {
-  has(data, 'type') && (data.type = data.type.toUpperCase());
-  has(data, 'config.size') && (data.width = data.config.size.toUpperCase());
-  has(data, 'config.alignment') && (data.alignment = data.config.alignment.toUpperCase());
+const convertDividerData = (data: {
+  type?: string;
+  config?: ComponentData['config'];
+  width;
+  alignment;
+  containerData;
+}) => {
+  has(data, 'type') && (data.type = data.type?.toUpperCase());
+  has(data, 'config.size') && (data.width = data.config?.size?.toUpperCase());
+  has(data, 'config.alignment') && (data.alignment = data.config?.alignment?.toUpperCase());
   data.containerData = { width: { type: PluginContainerData_Width_Type.CONTENT } };
 };
 
-const convertImageData = data => {
-  const { file_name, width, height } = data.src;
-  const { link, anchor, disableExpand } = data.config;
+const convertImageData = (data: {
+  src?: ImageComponentData;
+  config?: ComponentData['config'];
+  metadata?: { alt?: string; caption?: string };
+  image;
+  link;
+  disableExpand;
+  altText;
+  caption;
+}) => {
+  const { file_name, width, height } = data.src || {};
+  const { link, anchor, disableExpand } = data.config || {};
   data.image = { src: { custom: file_name }, width, height };
   data.link = (link || anchor) && convertLink({ ...link, anchor });
   data.disableExpand = disableExpand;
@@ -107,7 +133,7 @@ const convertImageData = data => {
   data.caption = data.metadata?.caption;
 };
 
-const convertPollData = data => {
+const convertPollData = (data: { layout; design }) => {
   has(data, 'layout.poll.type') && (data.layout.poll.type = data.layout.poll.type.toUpperCase());
   has(data, 'layout.poll.direction') &&
     (data.layout.poll.direction = data.layout.poll.direction.toUpperCase());
@@ -115,31 +141,44 @@ const convertPollData = data => {
     (data.design.poll.backgroundType = data.design.poll.backgroundType.toUpperCase());
 };
 
-const convertVerticalEmbedData = data => {
-  has(data, 'type') && (data.type = data.type.toUpperCase());
+const convertVerticalEmbedData = (data: { type?: string }) => {
+  has(data, 'type') && (data.type = data.type?.toUpperCase());
 };
 
-const convertLinkPreviewData = data => {
+const convertLinkPreviewData = (data: {
+  thumbnail_url?: string;
+  provider_url?: string;
+  config?: { link };
+  thumbnailUrl;
+  providerUrl;
+}) => {
   has(data, 'thumbnail_url') && (data.thumbnailUrl = data.thumbnail_url);
   has(data, 'provider_url') && (data.providerUrl = data.provider_url);
-  has(data, 'config.link') && (data.config.link = convertLink(data.config.link));
+  data.config?.link && (data.config.link = convertLink(data.config?.link));
 };
 
-const convertMention = data => {
-  data.name = data.mention.name;
-  data.slug = data.mention.slug;
+const convertMention = (data: {
+  mention?: { name?: string; slug?: string };
+  name?: string;
+  slug?: string;
+}) => {
+  data.name = data.mention?.name;
+  data.slug = data.mention?.slug;
   delete data.mention;
 };
 
-const convertFileData = data => {
+const convertFileData = (data: FileComponentData & { src }) => {
   const src: FileSource = { url: data.url, custom: data.id };
   data.src = src;
 };
 
-const convertButtonData = (data, blockType) => {
-  const { settings, design } = data.button;
-  const { borderRadius, borderWidth, background, color, borderColor } = design;
-  const { buttonText, url, rel, target } = settings;
+const convertButtonData = (
+  data: { button?: { settings; design }; styles; type; text; link },
+  blockType: string
+) => {
+  const { settings, design } = data.button || {};
+  const { borderRadius, borderWidth, background, color, borderColor } = design || {};
+  const { buttonText, url, rel, target } = settings || {};
   data.styles = {
     borderRadius,
     borderWidth,