+++ CHANGELOG.md
@@ -19,10 +19,21 @@
 ### :bug: Bug Fix
 - `editor`
   - [#2385](https://github.com/wix-incubator/rich-content/pull/2385) toolbar shortcuts in mac & windows
+### :house: Internal
+- `fullscreen`
+  - [#2392](https://github.com/wix/ricos/pull/2392) replace content-box with border-box
+- `docs`
+  - [#2337](https://github.com/wix-incubator/rich-content/pull/2337) docs removed from workspaces (build failure fix)
 
 </details>
 <hr/>
 
+## 8.30.3 (April 22, 2021)
+### :bug: Bug Fix
+- `image`
+  - [#2390](https://github.com/wix-incubator/rich-content/pull/2390) click on image caption doesnt open expand mode
+- `ricos-content`
+  - [#2393](https://github.com/wix-incubator/rich-content/pull/2393) draft converters handle undefined values in content
 ## 8.30.2 (April 22, 2021)
 ### :bug: Bug Fix
 - `editor`